+++ NEWS.md
@@ -1,8 +1,11 @@
 # News for Package CovTools
+### Changes in version 0.5.3
+  * Fix an error in `CovEst.2010OAS`.
+  
 ### Changes in version 0.5.2
   * Function list is moved to README file.
   * Following functions are added,
-    - `CovEst.2003LW`, `CovEst.2010OAS`, `CovEst.2010RBLW`
+    - `CovEst.2003LW`, `CovEst.2010OAS`, `CovEst.2010RBLW`.
     
 ### Changes in version 0.5.1
   * Functions for hypothesis tests are separated for clarity.